requirements.txt
langchain
langchain_ollama
python-dotenv
fastapi
uvicorn
pymysql
sqlalchemy
alembic
sqlmodel